Make suggestions change based on room
